requirements.txt
numpy
scipy
trimesh[easy]
ezdxf
networkx
PyYaml